+++ aat-lib/src/main/java/ch/bailu/aat_lib/map/layer/gpx/painter/RoutePainterNode.kt
@@ -10,10 +10,10 @@ class RoutePainterNode(private val mcontext: MapContext, private val color: Int,
     override fun drawEdge(nodes: TwoNodes) {}
     override fun drawNode(node: TwoNodes.PixelNode) {
         val c: Int
-        val altitude = node.point.getAltitude().toInt()
+        val altitude = node.point.getAltitude()
         c =
             if (altitude == ElevationProvider.NULL_ALTITUDE) color else AltitudeColorTable.instance()
-                .getColor(altitude)
+                .getColor(altitude.toInt())
         mcontext.draw().bitmap(mcontext.draw().getNodeBitmap(), node.pixel, c)
     }
 }
